Tweak some error strings
I noticed that when I was missing the composition-resource-name
annotation the error seemed to be missing something. This was because my
composed resource template didn't have a metadata.name, and the error
was trying to use the name to identify the resource.

It's pretty common to leave out the metadata.name and instead let
Crossplane generate one. So instead I've switched to using the Kind to
identify the resource. It's not quite as accurate, but it's always
required.

While I was in there I tweaked some of the other error strings to
provide as much detail as possible. I also removed the error constants
pattern, which we're planning to move away from in c/c.

Signed-off-by: Nic Cope <nicc@rk0n.org>