Update README and examples
This commit refactors the README to add a little more documentation on
how to use the function. It also tries to match the README style of the
other composition functions.

I've also broken the examples out into two directories. The Inline
example can be rendered using the Crossplane CLI. I've simplified that
one a bit to streamline doing that. The FileSystem example can't, so I've
left it as it was.

Signed-off-by: Nic Cope <nicc@rk0n.org>