Rename the Input type to GoTemplate
I just realised the examples were updated but the actual kind wasn't.
This doesn't matter very much because right now nothing validates the
kind, so you could use whatever apiVersion and kind you wanted as long
as the struct schemas matched. :)

This does fix the Marketplace entry though, which is a little confusing
because it lists the input kind as 'Input' when the examples here are
all 'GoTemplate'.

Signed-off-by: Nic Cope <nicc@rk0n.org>